# packages/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Room Service Hadith API
  version: 0.1.0
  description: |
    OpenAPI contract for the Hotel Guest Service Platform MVP.

    The API supports guest requests from a room-bound QR code or Smart TV,
    operational department dashboards, Room Manager monitoring, and
    Receptionist guest-room assignment. Staff authentication/RBAC, Smart TV
    provisioning, Guest context/catalog/request APIs, and their memory and
    PostgreSQL repository adapters are implemented in apps/api. Guest screens
    and later operational domains remain incremental implementation work.
servers:
  - url: http://127.0.0.1:3000/api/v1
    description: Local development
...
paths:
...
  /guest/context:
    get:
      tags: [Guest Context]
      operationId: getGuestContext
      summary: Resolve the current room-bound guest context
      description: |
        Resolves a room from either the opaque token extracted from a QR URL
        or the mapped Smart TV device credential. The client does not submit a
        room number.
      security:
        - GuestAccessToken: []
        - TvDeviceCredential: []
      responses:
        '200':
          description: Room-bound guest context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestContext'
        '401':
          $ref: '#/components/responses/GuestContextUnauthorized'
        '404':
          $ref: '#/components/responses/ContextNotFound'
  /guest/departments:
    get:
      tags: [Guest Catalog]
      operationId: listGuestDepartments
      summary: List guest-visible departments and units
      security:
        - GuestAccessToken: []
        - TvDeviceCredential: []
      responses:
        '200':
          description: Departments and units available to the guest.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentListResponse'
        '401':
          $ref: '#/components/responses/GuestContextUnauthorized'
        '404':
          $ref: '#/components/responses/ContextNotFound'
  /guest/menus:
    get:
      tags: [Guest Catalog]
      operationId: listGuestMenus
      summary: List active menu items and services
      description: Menu routing is derived from the selected item; the client does not choose a dashboard destination.
      security:
        - GuestAccessToken: []
        - TvDeviceCredential: []
      parameters:
        - $ref: '#/components/parameters/GuestUnitFilter'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Active menu items for the requested unit.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItemListResponse'
        '401':
          $ref: '#/components/responses/GuestContextUnauthorized'
        '404':
          $ref: '#/components/responses/ContextNotFound'
  /guest/menus/{menuItemId}:
    get:
      tags: [Guest Catalog]
      operationId: getGuestMenuItem
      summary: Get one active menu item or service
      security:
        - GuestAccessToken: []
        - TvDeviceCredential: []
      parameters:
        - $ref: '#/components/parameters/MenuItemId'
      responses:
        '200':
          description: Active menu item.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
        '401':
          $ref: '#/components/responses/GuestContextUnauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /guest/requests:
    get:
      tags: [Guest Requests]
      operationId: listGuestRequests
      summary: List requests for the current room context
      description: 'Returns requests belonging only to the active guest assignment in the room-bound context.'
      security:
        - GuestAccessToken: []
        - TvDeviceCredential: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/RequestStatusFilter'
      responses:
        '200':
          description: Requests belonging to the current room context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestRequestListResponse'
        '401':
          $ref: '#/components/responses/GuestContextUnauthorized'
        '404':
          $ref: '#/components/responses/ContextNotFound'
    post:
      tags: [Guest Requests]
      operationId: createGuestRequest
      summary: Submit a request or order
      description: |
        Creates a NEW request. All submitted menu items must resolve to the
        same unit so that backend routing is unambiguous. The room is derived
        from the QR or TV context, never from the request body.
      security:
        - GuestAccessToken: []
        - TvDeviceCredential: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGuestRequest'
            examples:
              restaurant:
                value:
                  clientRequestId: 2d8f0b6e-1f7b-4c27-91c5-93d6f8b9c8c0
                  items:
                    - menuItemId: 2cbbf8a2-fb66-4ae4-8b95-d3b37c2a7b10
                      quantity: 1
                      note: No spicy sauce
      responses:
        '201':
          description: Request created with status NEW.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestRequest'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/GuestContextUnauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/RequestConflict'
  /guest/requests/{requestId}:
    get:
      tags: [Guest Requests]
      operationId: getGuestRequest
      summary: Get one room-bound request and its status history
      security:
        - GuestAccessToken: []
        - TvDeviceCredential: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Request status and timestamps.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestRequest'
        '401':
          $ref: '#/components/responses/GuestContextUnauthorized'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /tv/context:
    get:
      tags: [TV]
      operationId: getTvContext
      summary: Get Smart TV room and welcome state
      description: Returns the room mapped to the TV device and the current personalized welcome state.
      security:
        - TvDeviceCredential: []
      responses:
        '200':
          description: Smart TV context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TvContext'
        '401':
          $ref: '#/components/responses/DeviceUnauthorized'
        '404':
          $ref: '#/components/responses/ContextNotFound'
  /tv/provisioning/start:
    post:
      tags: [TV]
      operationId: startTvProvisioning
      summary: Start one-time TV pairing
      description: Creates a short-lived pairing code for a newly installed TV. No room number or device credential is accepted from the TV.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartTvProvisioningRequest'
      responses:
        '200':
          description: Pairing code created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartTvProvisioningResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /tv/provisioning/claim:
    post:
      tags: [TV]
      operationId: claimTvProvisioning
      summary: Claim a paired TV credential
      description: Exchanges a pairing code after receptionist confirmation for the mapped TV credential.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimTvProvisioningRequest'
      responses:
        '200':
          description: TV credential issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimTvProvisioningResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/ResourceConflict'
        '410':
          $ref: '#/components/responses/PairingExpired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /receptionist/tv-devices:
    get:
      tags: [Receptionist]
      operationId: listTvDevices
      summary: List TV devices managed by the receptionist
      security:
        - StaffSession: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: roomId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TvProvisioningStatus'
      responses:
        '200':
          description: TV devices without credentials or pairing secrets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TvDeviceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /receptionist/tv-devices/pair:
    post:
      tags: [Receptionist]
      operationId: pairTvDevice
      summary: Map a pending TV to a room
      description: Receptionist-only operation that binds a short-lived TV pairing code to a room.
      security:
        - StaffSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PairTvDeviceRequest'
      responses:
        '200':
          description: TV device mapped to the room.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PairTvDeviceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
        '409':
          $ref: '#/components/responses/ResourceConflict'
        '410':
          $ref: '#/components/responses/PairingExpired'
  /receptionist/tv-devices/{deviceId}/revoke:
    post:
      tags: [Receptionist]
      operationId: revokeTvDevice
      summary: Revoke a TV device credential
      security:
        - StaffSession: []
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: TV device revoked.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /receptionist/tv-devices/{deviceId}/reset:
    post:
      tags: [Receptionist]
      operationId: resetTvDevice
      summary: Reset a TV device for re-pairing
      description: Revokes the current credential, clears the room mapping, and starts a fresh short-lived pairing session for the same TV installation.
      security:
        - StaffSession: []
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: TV device reset and ready to pair again.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetTvDeviceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
...
  /receptionist/rooms:
    get:
      tags: [Receptionist]
      operationId: listReceptionistRooms
      summary: List rooms and active guest assignments
      security:
        - StaffSession: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/RoomStatusFilter'
        - $ref: '#/components/parameters/RoomSearch'
      responses:
        '200':
          description: Room list for Receptionist dashboard.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceptionistRoomListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /receptionist/rooms/{roomId}/guest-access-token:
    get:
      tags: [Receptionist]
      operationId: getGuestAccessTokenStatus
      summary: Get the room QR access status
      description: Returns QR metadata only. The opaque token is never returned by this endpoint.
      security:
        - StaffSession: []
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '200':
          description: Current QR status for the room.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestQrRoomStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
    post:
      tags: [Receptionist]
      operationId: issueGuestAccessToken
      summary: Issue the room QR access URL
      description: |
        Creates or rotates the one opaque QR token for a room. The token is
        returned only inside the QR URL; it is valid only while the room has
        an active guest assignment.
      security:
        - StaffSession: []
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '200':
          description: QR URL generated for the room.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestAccessTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /receptionist/rooms/{roomId}/guest-access-token/revoke:
    post:
      tags: [Receptionist]
      operationId: revokeGuestAccessToken
      summary: Revoke the room QR access token
      description: Revokes the active QR token for the room. Revocation is recoverable by issuing a new token.
      security:
        - StaffSession: []
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '200':
          description: Current room QR token revocation result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevokeGuestAccessTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
  /receptionist/guest-access-tokens/batch:
    post:
      tags: [Receptionist]
      operationId: issueGuestAccessTokensBatch
      summary: Issue room QR access URLs in one transaction
      description: Rotates one opaque QR token for each requested room. The operation is atomic.
      security:
        - StaffSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchGuestAccessTokenRequest'
      responses:
        '200':
          description: QR URLs generated for all requested rooms.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchGuestAccessTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/ResourceNotFound'
...
components:
  securitySchemes:
    StaffSession:
      type: apiKey
      in: cookie
      name: room_service_session
      description: Secure HttpOnly staff session cookie. Staff users do not use a bearer token in this MVP.
    GuestAccessToken:
      type: apiKey
      in: header
      name: X-Guest-Access-Token
      description: Opaque token extracted from the room QR URL. It is not a room number and is scoped to a room context.
    TvDeviceCredential:
      type: apiKey
      in: header
      name: X-Device-Credential
      description: Credential for the Smart TV device mapped to a room.
  parameters:
    Page:
      name: page
      in: query
      description: One-based page number.
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      description: Number of records per page.
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    RequestStatusFilter:
      name: status
      in: query
      description: Filter by operational request status.
      required: false
      schema:
        $ref: '#/components/schemas/RequestStatus'
...
    GuestUnitFilter:
      name: unit
      in: query
      description: Guest menu unit to load. Lounge is returned by departments but is disabled until configured.
      required: true
      schema:
        $ref: '#/components/schemas/UnitCode'
...
    RoomStatusFilter:
      name: status
      in: query
      description: Filter by application room status.
      required: false
      schema:
        $ref: '#/components/schemas/RoomStatus'
...
    RequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    MenuItemId:
      name: menuItemId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    Unauthorized:
      description: Staff session is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    GuestContextUnauthorized:
      description: Guest access token or TV device credential is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    DeviceUnauthorized:
      description: Smart TV device credential is missing or invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    Forbidden:
      description: The authenticated role or permission cannot access this resource.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ResourceNotFound:
      description: The requested resource does not exist or is outside the caller's scope.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ContextNotFound:
      description: The QR token or device is not mapped to a valid room context.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    RequestConflict:
      description: Duplicate client request ID, unavailable menu item, or cross-unit request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    ResourceConflict:
      description: The requested resource change conflicts with existing data.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    TooManyRequests:
      description: Too many authentication attempts.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    PairingExpired:
      description: The one-time TV pairing code has expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
...
    DepartmentListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DepartmentSummary'
...
    GuestContext:
      type: object
      required: [room, roomStatus, welcome, availableUnits]
      properties:
        room:
          $ref: '#/components/schemas/RoomReference'
        roomStatus:
          $ref: '#/components/schemas/RoomStatus'
        welcome:
          $ref: '#/components/schemas/WelcomeState'
        availableUnits:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/UnitCode'
...
    TvProvisioningStatus:
      type: string
      enum: [PENDING, PAIRED, CLAIMED, REVOKED]
...
    TvDeviceListResponse:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TvDeviceAdmin'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    TvContext:
      type: object
      required: [device, roomStatus, welcome]
      properties:
        device:
          $ref: '#/components/schemas/TvDevice'
        roomStatus:
          $ref: '#/components/schemas/RoomStatus'
        welcome:
          $ref: '#/components/schemas/WelcomeState'
    StartTvProvisioningRequest:
      type: object
      required: [installationId, appVersion, deviceModel, androidApiLevel]
      properties:
        installationId:
          type: string
          minLength: 16
          maxLength: 128
        appVersion:
          type: string
          minLength: 1
          maxLength: 50
        deviceModel:
          type: string
          minLength: 1
          maxLength: 120
        androidApiLevel:
          type: integer
          minimum: 21
          maximum: 100
    StartTvProvisioningResponse:
      type: object
      required: [deviceId, deviceCode, pairingCode, expiresAt]
      properties:
        deviceId:
          type: string
          format: uuid
        deviceCode:
          type: string
          example: device_a83ks9
        pairingCode:
          type: string
          pattern: '^[0-9]{6}$'
        expiresAt:
          type: string
          format: date-time
    ClaimTvProvisioningRequest:
      type: object
      required: [pairingCode, installationId]
      properties:
        pairingCode:
          type: string
          pattern: '^[0-9]{6}$'
        installationId:
          type: string
          minLength: 16
          maxLength: 128
    ClaimTvProvisioningResponse:
      type: object
      required: [credential, device]
      properties:
        credential:
          type: string
          minLength: 16
        device:
          $ref: '#/components/schemas/TvDevice'
    PairTvDeviceRequest:
      type: object
      required: [pairingCode, roomId, roomNumber]
      properties:
        pairingCode:
          type: string
          pattern: '^[0-9]{6}$'
        roomId:
          type: string
          format: uuid
        roomNumber:
          type: string
          minLength: 1
          maxLength: 20
    PairTvDeviceResponse:
      type: object
      required: [device, pairedAt]
      properties:
        device:
          $ref: '#/components/schemas/TvDevice'
        pairedAt:
          type: string
          format: date-time
...
    MenuItem:
      type: object
      required:
        - id
        - unit
        - kind
        - name
        - localizedName
        - sortOrder
        - active
        - available
        - quantityAllowed
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        unit:
          $ref: '#/components/schemas/UnitCode'
        kind:
          $ref: '#/components/schemas/MenuItemKind'
        name:
          type: string
          example: Nasi Goreng
        localizedName:
          $ref: '#/components/schemas/LocalizedText'
        sortOrder:
          type: integer
          minimum: 0
          example: 1
        description:
          type: string
          nullable: true
        localizedDescription:
          type: object
          nullable: true
          allOf:
            - $ref: '#/components/schemas/LocalizedDescription'
        price:
          type: number
          format: double
          minimum: 0
          nullable: true
          example: 85000
        currency:
          type: string
          nullable: true
          minLength: 3
          maxLength: 3
          example: IDR
        durationMinutes:
          type: integer
          minimum: 1
          nullable: true
          example: 60
        imageMediaId:
          type: string
          format: uuid
          nullable: true
        active:
          type: boolean
          example: true
        available:
          type: boolean
          example: true
        quantityAllowed:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateGuestRequest:
      type: object
      required: [clientRequestId, items]
      properties:
        clientRequestId:
          type: string
          format: uuid
          description: Client-generated idempotency key. Must be unique for the room context.
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/CreateGuestRequestItem'
        guestNote:
          type: string
          maxLength: 1000
          nullable: true
          description: Optional guest note retained with the request; it is not a payment note.
...
    GuestRequest:
      type: object
      required:
        [
          id,
          clientRequestId,
          department,
          unit,
          items,
          status,
          requestedAt,
          statusHistory,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: string
          format: uuid
        clientRequestId:
          type: string
          format: uuid
        department:
          $ref: '#/components/schemas/DepartmentCode'
        unit:
          $ref: '#/components/schemas/UnitCode'
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/RequestItem'
        guestNote:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/RequestStatus'
        requestedAt:
          type: string
          format: date-time
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        statusHistory:
          type: array
          items:
            $ref: '#/components/schemas/StatusHistoryEntry'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    GuestRequestListResponse:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/GuestRequest'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...
    MenuItemListResponse:
      type: object
      required: [items, page, pageSize, total]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MenuItem'
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
...